add google play lib for ios

+++ plugins/googleplay/proj.ios/GooglePlay/GooglePlayGames.framework/Versions/A/Headers/GPGMultiplayerConfig.h
@@ -9,7 +9,7 @@
 
 @interface GPGMultiplayerConfig : NSObject
 
-@property(nonatomic, assign) int exclusiveBitMask;
+@property(nonatomic, assign) unsigned long long exclusiveBitMask;
 
 @property(nonatomic, retain) NSArray *invitedPlayerIds;
